backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
chromadb>=0.4.22
crewai>=0.11.0
langchain>=0.1.0
langchain-community>=0.0.12
langchain-ollama>=0.1.0
google-generativeai>=0.3.0
ollama>=0.1.7
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.1.4
GitPython>=3.1.40
motor>=3.3.2
pymongo>=4.6.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0